+++ swaylortift/swaylortift/Source/Util/LogFile.swift
@@ -46,11 +46,20 @@ public class LogFile
 	{
 		if let fileHandle = FileHandle(forWritingAtPath: filePath)
 		{
-			fileHandle.seekToEndOfFile()
-			if let data = content.data(using: encoding)
+			if #available(macCatalyst 13.0, *)
 			{
-				fileHandle.write(data)
-				fileHandle.closeFile()
+				fileHandle.seekToEndOfFile()
+				if let data = content.data(using: encoding)
+				{
+					fileHandle.write(data)
+					fileHandle.closeFile()
+				}
+			}
+			else
+			{
+				// Fallback on earlier versions
+				Swift.print("[ERROR] Operation not supported: fileHandle.write", terminator: Log.terminator)
+
 			}
 		}
 		else
@@ -76,7 +85,15 @@ public class LogFile
 	{
 		do
 		{
-			try FileManager.default.removeItem(atPath: filePath)
+			if #available(macCatalyst 13.0, *)
+			{
+				try FileManager.default.removeItem(atPath: filePath)
+			}
+			else
+			{
+				// Fallback on earlier versions
+				Swift.print("[ERROR] Operation not supported: FileManager.default.removeItem", terminator: Log.terminator)
+			}
 		}
 		catch
 		{
